requirements.txt
PyAudio~=0.2.11
numpy~=1.20.3
matplotlib~=3.4.2
scipy~=1.6.3
Pillow~=8.2.0
spotipy~=2.18.0
credentials~=1.1
requests~=2.25.1